requirements.txt
flask
flask-cors
pymongo
python-dotenv
selenium
webdriver-manager
gunicorn